:sparkles: Added several models and validator
Site:
A site refers to a an actual site I go to. (Youtube, twitter etc)
Topic:
A topic is the subjects a Site, Project or Post may talk about.
Not saved in databse and just a module since it doesn't need to be
instanced.
Array Validator:
Simple validator for psql array columns.